fix: bad routes
