+++ server/controllers/chat.gpt.controller.js
@@ -8,20 +8,6 @@ module.exports = ({ strapi }) => ({
       .getResponsefromChatGpt(ctx);
     ctx.send(data);
   },
-  async getConfig(ctx) {
-    const config = await strapi
-      .plugin("strapi-supergpt")
-      .service("superGptService")
-      .getConfig();
-    ctx.send(config);
-  },
-  async updateConfig(ctx) {
-    const config = await strapi
-      .plugin("strapi-supergpt")
-      .service("superGptService")
-      .updateConfig(ctx);
-    ctx.send(config);
-  },
   async createImage(ctx) {
     const config = await strapi
     .plugin("strapi-supergpt")
