+++ server/controllers/index.js
@@ -1,8 +1,10 @@
 "use strict";
-const chatGptController = require("./chat-gpt.controller");
+const chatGptController = require("./chat.gpt.controller");
 const convoController = require('./convo.controller');
+const cacheController = require('./cache.controller')
 
 module.exports = {
+  cacheController,
   chatGptController,
   convoController
 };
