+++ server/routes/cache.routes.js
@@ -1,0 +1,22 @@
+"use strict";
+
+module.exports = {
+  routes: [
+    {
+      method: "GET",
+      path: "/cache",
+      handler: "cacheController.getConfig",
+      config: {
+        policies: [],
+      },
+    },
+    {
+      method: "POST",
+      path: "/cache/update",
+      handler: "cacheController.updateConfig",
+      config: {
+        policies: [],
+      },
+    },
+  ],
+};
