+++ server/routes/chat.gpt.routes.js
@@ -17,22 +17,6 @@ module.exports = {
       config: {
         policies: [],
       }
-    },
-    {
-      method: "GET",
-      path: "/config",
-      handler: "chatGptController.getConfig",
-      config: {
-        policies: [],
-      },
-    },
-    {
-      method: "POST",
-      path: "/config/update",
-      handler: "chatGptController.updateConfig",
-      config: {
-        policies: [],
-      },
-    },
+    }
   ],
 };
