+++ server/services/cache.service.js
@@ -1,0 +1,53 @@
+module.exports = ({strapi}) => ({
+	getConfig() {
+		try {
+		  const pluginStore = strapi.store({
+			environment: strapi.config.environment,
+			type: "plugin",
+			name: "strapi-supergpt",
+		  });
+
+		  return pluginStore.get({ key: "superGPTConfig" });
+		} catch (error) {
+		  strapi.log.error(error.message);
+		  return {
+			error:
+			  "An error occurred while fetching chatGPT config. Please try after some time",
+		  };
+		}
+	  },
+
+	  updateConfig(ctx) {
+		try {
+		  const reqBody = ctx.request.body;
+		  const data = {
+			apiKey: reqBody.apiKey,
+			modelName: reqBody.modelName || "gpt-3.5-turbo",
+			aiImageModelName: reqBody.aiImageModelName || "dall-e-3",
+			temperature: reqBody.temperature || 0.0,
+			maxTokens: reqBody.maxTokens || 2048,
+			topP: reqBody.topP,
+			frequencyPenalty: reqBody.frequencyPenalty || 0.0,
+			presencePenalty: reqBody.presencePenalty || 0.0,
+			stop: reqBody.stop || "",
+			convoCount: reqBody.Count || "",
+		  };
+		  const pluginStore = strapi.store({
+			environment: strapi.config.environment,
+			type: "plugin",
+			name: "strapi-supergpt",
+		  });
+
+		  return pluginStore.set({
+			key: "superGPTConfig",
+			value: data,
+		  });
+		} catch (error) {
+		  strapi.log.error(error.message);
+		  return {
+			error:
+			  "An error occurred while updting the chatGPT config. Please try after some time",
+		  };
+		}
+	  },
+})
