+++ server/services/chat-gpt.service.js
@@ -2,11 +2,14 @@
 const { OpenAI } = require("openai");
 const utils = require('../utils')
 
+
 module.exports = ({ strapi }) => ({
+  config: strapi.plugin("strapi-supergpt").service("cacheService"),
+
   async getResponsefromChatGpt(ctx) {
-    const config = await this.getConfig();
+    this.config.getConfig();
     const openai = new OpenAI({
-      apiKey: config.apiKey,
+      apiKey: this.config.apiKey,
     });
 
     const {
@@ -21,8 +24,8 @@ module.exports = ({ strapi }) => ({
     } = ctx.request.body;
     try {
       const requestParams = {
-        model: config.modelName,
-        max_tokens: config.maxTokens ? parseInt(config.maxTokens) : 2048,
+        model: this.config.modelName,
+        max_tokens: this.config.maxTokens ? parseInt(this.config.maxTokens) : 2048,
         prompt: prompt.trim(),
       };
 
@@ -93,57 +96,6 @@ module.exports = ({ strapi }) => ({
       return {
         error:
           "An error occurred while fetching the chat response. Please try after some time",
-      };
-    }
-  },
-
-  getConfig() {
-    try {
-      const pluginStore = strapi.store({
-        environment: strapi.config.environment,
-        type: "plugin",
-        name: "strapi-supergpt",
-      });
-
-      return pluginStore.get({ key: "superGPTConfig" });
-    } catch (error) {
-      strapi.log.error(error.message);
-      return {
-        error:
-          "An error occurred while fetching chatGPT config. Please try after some time",
-      };
-    }
-  },
-
-  updateConfig(ctx) {
-    try {
-      const reqBody = ctx.request.body;
-      const data = {
-        apiKey: reqBody.apiKey,
-        modelName: reqBody.modelName || "gpt-3.5-turbo",
-        aiImageModelName: reqBody.aiImageModelName || "dall-e-3",
-        temperature: reqBody.temperature || 0.0,
-        maxTokens: reqBody.maxTokens || 2048,
-        topP: reqBody.topP,
-        frequencyPenalty: reqBody.frequencyPenalty || 0.0,
-        presencePenalty: reqBody.presencePenalty || 0.0,
-        stop: reqBody.stop || "",
-      };
-      const pluginStore = strapi.store({
-        environment: strapi.config.environment,
-        type: "plugin",
-        name: "strapi-supergpt",
-      });
-
-      return pluginStore.set({
-        key: "superGPTConfig",
-        value: data,
-      });
-    } catch (error) {
-      strapi.log.error(error.message);
-      return {
-        error:
-          "An error occurred while updting the chatGPT config. Please try after some time",
       };
     }
   },
