+++ server/services/index.js
@@ -2,8 +2,10 @@
 
 const superGptService = require("./chat-gpt.service");
 const convoService = require('./convo.service')
+const cacheService = require('./cache.service')
 
 module.exports = {
   superGptService,
-  convoService
+  convoService,
+  cacheService
 };
